plug matomo to get basic page update
