+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import { Routes, Route, useLocation } from "react-router-dom";
+import { useMatomo } from "@jonkoops/matomo-tracker-react";
 import axios from "axios";
 import {
   EcoScorePage,
@@ -20,32 +20,6 @@ import DevModeContext from "./contexts/devMode";
 import { getIsDevMode, getVisiblePages } from "./localeStorageManager";
 import LoginContext from "./contexts/login";
 import off from "./off";
-
-const theme = createTheme({
-  palette: {
-    success: {
-      main: "#8bc34a",
-      contrastText: "#ffffff",
-    },
-    error: {
-      main: "#ff5252",
-      contrastText: "#ffffff",
-    },
-    primary: {
-      main: "#ff8714",
-      contrastText: "#ffffff",
-    },
-    secondary: {
-      main: "#ffefb7",
-      contrastText: "rgba(0, 0, 0, 0.87)",
-    },
-  },
-  components: {
-    MuiLink: {
-      defaultProps: { color: "inherit" },
-    },
-  },
-});
 
 export default function App() {
   const [devMode, setDevMode] = React.useState(getIsDevMode);
@@ -103,35 +77,38 @@ export default function App() {
     firstFetch();
   });
 
+  const location = useLocation();
+  const { trackPageView } = useMatomo();
+
+  React.useEffect(() => {
+    trackPageView();
+  }, [location, trackPageView]);
+
   return (
-    <Router>
-      <ThemeProvider theme={theme}>
-        <LoginContext.Provider value={{ ...userState, refresh }}>
-          <DevModeContext.Provider
-            value={{
-              devMode,
-              setDevMode,
-              visiblePages,
-              setVisiblePages,
-            }}
-          >
-            <CssBaseline />
-            <ResponsiveAppBar />
-            <Routes>
-              <Route path="/" element={<Home />} />
-              <Route path="/eco-score" element={<EcoScorePage />} />
-              <Route path="/logos" element={<LogoAnnotationPage />} />
-              <Route path="/logos/search" element={<LogoSearchPage />} />
-              <Route path="/logos/:logoId" element={<LogoUpdatePage />} />
-              <Route path="/settings" element={<SettingsPage />} />
-              <Route path="/questions" element={<QuestionsPage />} />
-              <Route path="/insights" element={<InsightsPage />} />
-              <Route path="*" element={<NotFoundPage />} />
-              <Route path="/nutriscore" element={<NutriscoreValidator />} />
-            </Routes>
-          </DevModeContext.Provider>
-        </LoginContext.Provider>
-      </ThemeProvider>
-    </Router>
+    <LoginContext.Provider value={{ ...userState, refresh }}>
+      <DevModeContext.Provider
+        value={{
+          devMode,
+          setDevMode,
+          visiblePages,
+          setVisiblePages,
+        }}
+      >
+        <CssBaseline />
+        <ResponsiveAppBar />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/eco-score" element={<EcoScorePage />} />
+          <Route path="/logos" element={<LogoAnnotationPage />} />
+          <Route path="/logos/search" element={<LogoSearchPage />} />
+          <Route path="/logos/:logoId" element={<LogoUpdatePage />} />
+          <Route path="/settings" element={<SettingsPage />} />
+          <Route path="/questions" element={<QuestionsPage />} />
+          <Route path="/insights" element={<InsightsPage />} />
+          <Route path="*" element={<NotFoundPage />} />
+          <Route path="/nutriscore" element={<NutriscoreValidator />} />
+        </Routes>
+      </DevModeContext.Provider>
+    </LoginContext.Provider>
   );
 }
